Revert "Add silk fibers aware socket implementations"
