+++ base/poco/Net/include/Poco/Net/SocketImpl.h
@@ -215,7 +215,7 @@ namespace Net
         /// Returns true if the next operation corresponding to
         /// mode will not block, false otherwise.
 
-        virtual bool pollImpl(Poco::Timespan & timeout, int mode);
+        bool pollImpl(Poco::Timespan & timeout, int mode);
         /// Modifies `timeout`
 
         virtual void setSendBufferSize(int size);
@@ -386,10 +386,6 @@ namespace Net
         virtual bool secure() const;
         /// Returns true iff the socket's connection is secure
         /// (using SSL or TLS).
-
-        virtual bool supportsExternalPolling() const;
-        /// Returns true iff an external poller can drive this socket's
-        /// readiness (the non-blocking / EAGAIN model).
 
         int socketError();
         /// Returns the value of the SO_ERROR socket option.
