+++ base/poco/Net/src/SocketImpl.cpp
@@ -496,12 +496,6 @@ bool SocketImpl::secure() const
 }
 
 
-bool SocketImpl::supportsExternalPolling() const
-{
-	return true;
-}
-
-
 bool SocketImpl::pollImpl(Poco::Timespan& remainingTime, int mode)
 {
 	poco_socket_t sockfd = _sockfd;
