+++ base/poco/NetSSL_OpenSSL/include/Poco/Net/SecureStreamSocketImpl.h
@@ -232,11 +232,6 @@ namespace Net
         /// This method will only work if the blocking modes of
         /// the socket are changed via the setBlocking method!
 
-        void setBioMethod(const BIO_METHOD * method);
-        /// Optionally inject a custom BIO_METHOD into the underlying SSL machinery.
-        /// Has no effect once the SSL handshake has been initiated (i.e. once
-        /// any I/O has happened). If never called, `BIO_s_socket()` is used.
-
 
     protected:
         void acceptSSL();
