+++ base/poco/NetSSL_OpenSSL/src/SecureSocketImpl.cpp
@@ -98,29 +98,14 @@ SocketImpl* SecureSocketImpl::acceptConnection(SocketAddress& clientAddr)
 }
 
 
-void SecureSocketImpl::setBioMethod(const BIO_METHOD * method)
-{
-	_bioMethod = method;
-}
-
-
-const BIO_METHOD * SecureSocketImpl::getBioMethod() const
-{
-	return _bioMethod ? _bioMethod : BIO_s_socket();
-}
-
-
 void SecureSocketImpl::acceptSSL()
 {
 	std::lock_guard<std::recursive_mutex> lock(_mutex);
 	poco_assert (!_pSSL);
 
-	BIO* pBIO = BIO_new(getBioMethod());
+	BIO* pBIO = BIO_new(BIO_s_socket());
 	if (!pBIO) throw SSLException("Cannot create BIO object");
 	BIO_set_fd(pBIO, static_cast<int>(_pSocket->sockfd()), BIO_NOCLOSE);
-
-	if (_bioMethod)
-		BIO_set_data(pBIO, _pSocket.get());
 
 	_pSSL = SSL_new(_pContext->sslContext());
 	if (!_pSSL)
@@ -184,12 +169,9 @@ void SecureSocketImpl::connectSSL(bool performHandshake)
 	poco_assert (!_pSSL);
 	poco_assert (_pSocket->initialized());
 
-	BIO* pBIO = BIO_new(getBioMethod());
+	BIO* pBIO = BIO_new(BIO_s_socket());
 	if (!pBIO) throw SSLException("Cannot create SSL BIO object");
 	BIO_set_fd(pBIO, static_cast<int>(_pSocket->sockfd()), BIO_NOCLOSE);
-
-	if (_bioMethod)
-		BIO_set_data(pBIO, _pSocket.get());
 
 	_pSSL = SSL_new(_pContext->sslContext());
 	if (!_pSSL)
