+++ src/IO/WriteBufferFromPocoSocket.cpp
@@ -197,12 +197,4 @@ WriteBufferFromPocoSocket::WriteBufferFromPocoSocket(Poco::Net::Socket & socket_
     write_event = write_event_;
 }
 
-void WriteBufferFromPocoSocket::setAsyncCallback(AsyncCallback async_callback_)
-{
-    if (async_callback_ && !socket.impl()->supportsExternalPolling())
-        throw Exception(ErrorCodes::LOGICAL_ERROR,
-            "Cannot set an async callback on a socket that does not support external polling");
-    async_callback = std::move(async_callback_);
-}
-
 }
