+++ src/IO/WriteBufferFromPocoSocket.h
@@ -19,7 +19,7 @@ public:
     explicit WriteBufferFromPocoSocket(Poco::Net::Socket & socket_, size_t buf_size = DBMS_DEFAULT_BUFFER_SIZE);
     explicit WriteBufferFromPocoSocket(Poco::Net::Socket & socket_, const ProfileEvents::Event & write_event_, size_t buf_size = DBMS_DEFAULT_BUFFER_SIZE);
 
-    void setAsyncCallback(AsyncCallback async_callback_);
+    void setAsyncCallback(AsyncCallback async_callback_) { async_callback = std::move(async_callback_); }
 
     using WriteBuffer::write;
     void write(const std::string & str) { WriteBuffer::write(str.c_str(), str.size()); }
