+++ src/configure_config.cmake
@@ -155,9 +155,6 @@ endif()
 if (TARGET ch_contrib::liburing)
     set(USE_LIBURING 1)
 endif ()
-if (TARGET ch_contrib::silk_fibers)
-    set(USE_SILK 1)
-endif ()
 if (TARGET ch_contrib::protobuf)
     set(USE_PROTOBUF 1)
 endif()
